Extract `getDynamicMeta`
